Small changes to mark for user extension
- Rename authoring actions and don't include it on article entity
- Include superdesk config type on superdesk-api
