+++ scripts/extensions/markForUser/src/get-authoring-actions.tsx
@@ -1,10 +1,10 @@
 import {ISuperdesk, IArticle, IArticleAction} from 'superdesk-api';
 import {manageMarkedUserForSingleArticle} from './managed-marked-user';
 
-export function getActionsExtraInitialize(superdesk: ISuperdesk) {
+export function authoringActionsInitialize(superdesk: ISuperdesk) {
     const {gettext} = superdesk.localization;
 
-    return function getActionsExtra(article: IArticle) {
+    return function getAuthoringActions(article: IArticle) {
         const markForUser: IArticleAction = {
             label: gettext('User'),
             onTrigger: () => manageMarkedUserForSingleArticle(superdesk, article),
